zkteco-desktop/backend/requirements.txt
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
future==0.18.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
packaging==23.2
python-dotenv==1.0.0
-e git+https://github.com/zeidanbm/pyzk.git@9cd5731543e3839a94962403c7ad7a5e9c872bac#egg=pyzk
referencing==0.32.0
prettytable
requests==2.31.0
rpds-py==0.15.2
sentry-sdk==1.39.2
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
flask-cors==4.0.0
psutil==5.9.6
pyinstaller==6.9.0
APScheduler==3.10.4
